Front work

+++ src/main/java/com/example/isaprojekat/service/impl/JpaBoatService.java
@@ -1,6 +1,7 @@
 package com.example.isaprojekat.service.impl;
 
 import com.example.isaprojekat.model.Boat;
+import com.example.isaprojekat.model.House;
 import com.example.isaprojekat.repository.BoatRepository;
 import com.example.isaprojekat.service.BoatService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,7 @@ import java.util.Optional;
 public class JpaBoatService implements BoatService {
     @Autowired
     private BoatRepository boatRepository;
+
     @Override
     public Optional<Boat> findOne(Long id) {
         return boatRepository.findById(id);
@@ -29,7 +31,38 @@ public class JpaBoatService implements BoatService {
 
     @Override
     public Boolean delete(Long id) {
-        boatRepository.deleteById(id);
-        return true;
+        Optional<Boat> b = boatRepository.findById(id);
+        if(b.isPresent()) {
+            boatRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
+    @Override
+    public Boat update(Boat boat) {
+        return boatRepository.save(boat);
+    }
+
+    @Override
+    public List<Boat> find(String name, String address, Double price) {
+        if(price == null){
+            price = Double.MAX_VALUE;
+        }
+
+        if(name != null && address == null){
+            return boatRepository.findByNameContainingAndPriceLessThanEqual(name, price);
+        }
+        else if(name != null && address != null){
+            return boatRepository.findByNameAndAddressContainingAndPriceLessThanEqual(name, address, price);
+        }
+        else if (name ==null && address != null){
+            return boatRepository.findByAddressContainingAndPriceLessThanEqual(address, price);
+        }
+
+        return boatRepository.findByPriceLessThanEqual(price);
+
+    }
+
+
 }
